Move HistoryManager, HistoryRecord to history package
