+++ src/main/java/cs/f10/t1/nursetraverse/logic/commands/HistoryCommand.java
@@ -3,7 +3,7 @@ package cs.f10.t1.nursetraverse.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import cs.f10.t1.nursetraverse.commons.util.CollectionUtil;
-import cs.f10.t1.nursetraverse.model.HistoryRecord;
+import cs.f10.t1.nursetraverse.model.history.HistoryRecord;
 import cs.f10.t1.nursetraverse.model.Model;
 import javafx.collections.ObservableList;
